go v4 project structure changes reverted.
