+++ controllers/devfileregistrieslist_controller_test.go
@@ -14,7 +14,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package controller
+package controllers
 
 import (
 	"fmt"
@@ -26,61 +26,62 @@ import (
 	"k8s.io/apimachinery/pkg/types"
 )
 
-var _ = Describe("ClusterDevfileRegistriesList controller test", func() {
-	Context("When Creating ClusterDevfileRegistriesList CR with valid values", func() {
+var _ = Describe("DevfileRegistriesList controller test", func() {
+	Context("When Creating DevfileRegistriesList CR with valid values", func() {
 		It("Should return a status saying URL is reachable", func() {
-			Expect(k8sClient.Create(ctx, getClusterDevfileRegistriesListCR(clusterdevfileRegistriesListName, devfileRegistriesNamespace,
+			Expect(k8sClient.Create(ctx, getDevfileRegistriesListCR(devfileRegistriesListName, devfileRegistriesNamespace,
 				devfileStagingRegistryName, devfileStagingRegistryURL))).Should(Succeed())
-			drlLookupKey := types.NamespacedName{Name: clusterdevfileRegistriesListName, Namespace: devfileRegistriesNamespace}
-			//validate success status
-			validateStatus(drlLookupKey, ClusterListType, allRegistriesReachable)
+			drlLookupKey := types.NamespacedName{Name: devfileRegistriesListName, Namespace: devfileRegistriesNamespace}
+			// validate success status
+			validateStatus(drlLookupKey, NamespaceListType, allRegistriesReachable)
 			//delete all crs
-			deleteCRList(drlLookupKey, ClusterListType)
+			deleteCRList(drlLookupKey, NamespaceListType)
 		})
 	})
-	Context("When Devfile URL is unavailable", func() {
-		It("Should return a status saying URL cannot be reached ", func() {
+	Context("When Updating DevfileRegistriesList CR with valid values", func() {
+		It("Should return a status saying all URLs are reachable ", func() {
 			//start mock index server
 			testServer := GetNewUnstartedTestServer()
 			Expect(testServer).ShouldNot(BeNil())
 			testServer.Start()
-			Expect(k8sClient.Create(ctx, getClusterDevfileRegistriesListCR(clusterdevfileRegistriesListName, devfileRegistriesNamespace,
+			Expect(k8sClient.Create(ctx, getDevfileRegistriesListCR(devfileRegistriesListName, devfileRegistriesNamespace,
 				localRegistryName, testServer.URL))).Should(Succeed())
 
-			drlLookupKey := types.NamespacedName{Name: clusterdevfileRegistriesListName, Namespace: devfileRegistriesNamespace}
+			drlLookupKey := types.NamespacedName{Name: devfileRegistriesListName, Namespace: devfileRegistriesNamespace}
 			// validate success status
-			validateStatus(drlLookupKey, ClusterListType, allRegistriesReachable)
+			validateStatus(drlLookupKey, NamespaceListType, allRegistriesReachable)
 
 			//shut down the mock server
 			testServer.Close()
 
 			//update the CR to trigger a reconcile
-			drl := &v1alpha1.ClusterDevfileRegistriesList{}
+			drl := &v1alpha1.DevfileRegistriesList{}
 			k8sClient.Get(ctx, drlLookupKey, drl)
 			//update list in existing CR
 			registriesList := drl.Spec.DevfileRegistries
 			registriesList = append(registriesList, v1alpha1.DevfileRegistryService{Name: devfileStagingRegistryName, URL: devfileStagingRegistryURL})
 			drl.Spec.DevfileRegistries = registriesList
 			Expect(k8sClient.Update(ctx, drl)).Should(Succeed())
-			// validate unreachable status
-			validateStatus(drlLookupKey, ClusterListType, fmt.Sprintf(registryUnreachable, testServer.URL))
+			// verify unreachable status
+			validateStatus(drlLookupKey, NamespaceListType, fmt.Sprintf(registryUnreachable, testServer.URL))
 
 		})
 	})
+
 	Context("When all entries in DevfileRegistry service list is deleted", func() {
 		It("Should return a status saying CR is empty", func() {
-			drlLookupKey := types.NamespacedName{Name: clusterdevfileRegistriesListName, Namespace: devfileRegistriesNamespace}
+			drlLookupKey := types.NamespacedName{Name: devfileRegistriesListName, Namespace: devfileRegistriesNamespace}
 
-			drl := &v1alpha1.ClusterDevfileRegistriesList{}
+			drl := &v1alpha1.DevfileRegistriesList{}
 			k8sClient.Get(ctx, drlLookupKey, drl)
 			registriesList := drl.Spec.DevfileRegistries
 			//delete all entries in the registries list
 			drl.Spec.DevfileRegistries = registriesList[:0]
 			Expect(k8sClient.Update(ctx, drl)).Should(Succeed())
 			//validate empty status
-			validateStatus(drlLookupKey, ClusterListType, emptyStatus)
+			validateStatus(drlLookupKey, NamespaceListType, emptyStatus)
 			//delete all crs
-			deleteCRList(drlLookupKey, ClusterListType)
+			deleteCRList(drlLookupKey, NamespaceListType)
 		})
 	})
